ARM: dts: bcm2711-rpi-400: Restore the ACT LED
Pi 400 doesn't have an ACT LED, only a green PWR LED, but a user may
wish to remap the ACT functionality to a GPIO on the 40-pin header.
Restore the ACT LED declaration, but leave it disabled in order to
prevent a GPIO being claimed until either the act_led_gpio dtparam
or the act-led overlay is used.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>